python/requirements.txt
enum34==1.1.6
grpcio==1.0.4
grpcio-tools==1.0.4
protobuf==3.1.0.post1
py==1.4.32
pytest==3.0.5
six==1.10.0
